pyxylookup/requirements.txt
requests==2.18.4
setuptools==38.4.0
msgpack==0.5.6
pandas==0.22.0
vcrpy==1.11.1
numpy==1.14.0
nose==1.3.7
